cv-parse/pyproject.toml
[project]
name = "cv-parse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "azure-storage-blob>=12.25.0",
    "azure-data-tables>=12.7.0",
    "azure-identity>=1.22.0",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.59",
    "langchain-openai>=0.2.14",
    "langchain-tools>=0.1.34",
    "llama-parse==0.4.9",
    "passlib[bcrypt]>=1.7.4",
    "psutil>=7.0.0",
    "pydantic[email]>=2.11.4",
    "pymupdf==1.26.1",
    "pyodbc>=5.2.0",
    "pypdf2>=3.0.1",
    "pytest-asyncio>=0.26.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy==2.0.40",
    "langgraph>=0.5.0",
    "uvicorn>=0.34.2",
    "pdfminer>=20191125",
    "pdfminer-six>=20250506",
    "python-dateutil>=2.9.0.post0",
]

[project.scripts]
kickoff = "cv_parse_flows.main:kickoff"
plot = "cv_parse_flows.main:plot"
cv-rank-flow = "cv_parse_flows.main:main"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cv_parse"]
